requirements.txt
python-dotenv
youtube-transcript-api
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-openai
faiss-cpu
